policyfile: support ETag parameters with escaped double quotes
Changes Set and SetAndGet for ACLs to support ETag parameters that are
passed as strings enclosed in escaped double quotes, e.g. "xyzzy".

The default ETag format as per the spec includes the double quotes and the
API returns ETags in double quotes, however the client currently rejects
ETag parameters that come in double quotes.

Changing this streamlines the experience of users who want to rely on the
ETag in scenarios where they read, modify and write an ACL with the client
and want to ensure the ACL hasn’t been modified between reading and writing.

Fixes #27

Signed-off-by: Gesa Stupperich <gesa@tailscale.com>